# services/orchestrator-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karaoke-Inator Orchestrator",
    "description": "Master orchestration service for YouTube audio processing: voice separation and lyrics extraction",
    "version": "1.0.0"
  },
  "paths": {
    "/process/queue": {
      "post": {
        "tags": [
          "Processing"
        ],
        "summary": "Queue audio processing from YouTube URL",
        "operationId": "queue_audio_processing_process_queue_post",
        "parameters": [
          {
            "name": "youtube_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Youtube Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/process/search": {
      "get": {
        "tags": [
          "Processing"
        ],
        "summary": "Search processing results by lyrics query",
        "operationId": "search_results_by_query_process_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/process/processing/{id}": {
      "get": {
        "tags": [
          "Processing"
        ],
        "summary": "Get Processing Result By Id",
        "operationId": "get_processing_result_by_id_process_processing__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/process/queue/position/{task_id}": {
      "get": {
        "tags": [
          "Processing"
        ],
        "summary": "Get the position of a task in the Redis queue",
        "operationId": "get_queue_position_process_queue_position__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuePositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/process/job/{job_id}/latest-message": {
      "get": {
        "tags": [
          "Processing"
        ],
        "summary": "Get the latest message from a job's progress stream for reconnection",
        "operationId": "get_latest_job_message_process_job__job_id__latest_message_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Chunk": {
        "properties": {
          "start": {
            "type": "number",
            "title": "Start",
            "description": "Start time of the segment in seconds",
            "example": 0.0
          },
          "end": {
            "type": "number",
            "title": "End",
            "description": "End time of the segment in seconds",
            "example": 12.92
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Lyrics text for this segment",
            "example": "Wake up in the mornin' feelin' like P. Diddy"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end",
          "text"
        ],
        "title": "Chunk"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobResponse"
      },
      "LatestMessageResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingOutputMessage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "message"
        ],
        "title": "LatestMessageResponse"
      },
      "Lyrics": {
        "properties": {
          "processing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Processing Id"
          },
          "full_text": {
            "type": "string",
            "title": "Full Text"
          },
          "chunks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Chunk"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunks"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          }
        },
        "type": "object",
        "required": [
          "processing_id",
          "full_text"
        ],
        "title": "Lyrics"
      },
...
      "ProcessingOutputMessage": {
        "properties": {
          "action": {
            "type": "string",
            "const": "processing_output",
            "title": "Action",
            "default": "processing_output"
          },
          "payload": {
            "$ref": "#/components/schemas/ProcessingOutputPayload"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "ProcessingOutputMessage"
      },
      "ProcessingOutputPayload": {
        "properties": {
          "current_step": {
            "type": "integer",
            "title": "Current Step",
            "description": "Current step number (0-5)",
            "example": 2
          },
          "total_steps": {
            "type": "integer",
            "title": "Total Steps",
            "description": "Total number of steps",
            "default": 5,
            "example": 5
          },
          "desc": {
            "type": "string",
            "title": "Desc",
            "description": "Progress or status message",
            "example": "Downloading audio"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "Result information (only set on completion or error)"
          }
        },
        "type": "object",
        "required": [
          "current_step",
          "desc"
        ],
        "title": "ProcessingOutputPayload"
      },
      "ProcessingResponse": {
        "properties": {
          "youtube_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Url"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "uploader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploader"
          },
          "uploader_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploader Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "vocals_minio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocals Minio Path"
          },
          "accompaniment_minio_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accompaniment Minio Path"
          },
          "lyrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Lyrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ProcessingResponse"
      },
      "ProcessingResult": {
        "properties": {
          "success": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success",
            "description": "Whether processing was successful (only set on completion)",
            "example": true
          },
          "vocals_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocals Path",
            "description": "Download URL for extracted vocals track",
            "example": "/download/vocals.mp3"
          },
          "accompaniment_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accompaniment Path",
            "description": "Download URL for accompaniment track",
            "example": "/download/accompaniment.mp3"
          },
          "lyrics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lyrics",
            "description": "Full transcribed lyrics as plain text",
            "example": "Never gonna give you up\nNever gonna let you down"
          },
          "chunks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Chunk"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunks",
            "description": "List of lyrics segments with timestamps"
          },
          "yt_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/YouTubeVideoMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "YouTube video metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if processing failed",
            "example": "Invalid YouTube URL"
          }
        },
        "type": "object",
        "title": "ProcessingResult"
      },
      "QueuePositionResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "total_in_queue": {
            "type": "integer",
            "title": "Total In Queue"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "position",
          "total_in_queue"
        ],
        "title": "QueuePositionResponse"
      },
      "SearchResultsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessingResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SearchResultsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "YouTubeVideoMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "uploader": {
            "type": "string",
            "title": "Uploader"
          },
          "uploader_url": {
            "type": "string",
            "title": "Uploader Url"
          },
          "thumbnail": {
            "type": "string",
            "title": "Thumbnail"
          },
          "thumbnail_url": {
            "type": "string",
            "title": "Thumbnail Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "uploader_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploader Id"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          },
          "like_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Like Count"
          },
          "dislike_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dislike Count"
          },
          "upload_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Date"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "age_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Limit"
          },
          "is_live": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Live"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "channel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Url"
          },
          "duration_string": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration String"
          },
          "format_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Id"
          },
          "format_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Note"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "vcodec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vcodec"
          },
          "acodec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acodec"
          },
          "extent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extent"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "title",
          "uploader",
          "uploader_url",
          "thumbnail",
          "thumbnail_url"
        ],
        "title": "YouTubeVideoMetadata"
      }
    }
  }
}